requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pytesseract==0.3.13
python-telegram-bot==21.9
pyzbar==0.1.9
pillow==11.0.0
python-multipart==0.0.20
